% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_dic.R
\name{calculate_dic}
\alias{calculate_dic}
\title{Calculate dissolved inorganic carbon}
\usage{
calculate_dic(ph, alkalinity)
}
\arguments{
\item{ph}{pH at which to perform the calculation}

\item{alkalinity}{Alkalinity, in mg CaCO3/L}
}
\value{

}
\description{
Calculate dissolved inorganic carbon
}
\examples{
calculate_dic(7.5, 20)
}
