% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_wdensity.R
\name{calculate_wdensity}
\alias{calculate_wdensity}
\title{Calculate the density of air-free water at 0 - 150 C and 101.325 kPa}
\usage{
calculate_wdensity(temp)
}
\arguments{
\item{temp}{A vector of water temperatures, in degrees celsius.}
}
\value{
A vector of densities, in kg/L.
}
\description{
Calculate the density of air-free water at 0 - 150 C and 101.325 kPa
}
\examples{
calculate_wdensity(25)
}
