% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_logk.R
\name{change_logk}
\alias{change_logk}
\title{Change log K values in a database}
\usage{
change_logk(db = pbcusol:::pbcu2sol, eqn, logk)
}
\arguments{
\item{db}{The PHREEQC database to modify.}

\item{eqn}{A character vector of equations to modify.}

\item{logk}{A character vector containing the replacement log K values, corresponding element-wise to \code{eqn}.}
}
\value{
A modified PHREEQC database.
}
\description{
Change log K values in a database
}
\examples{
phreeqc::minteq.v4.dat[1820:1821]
change_logk(db = phreeqc::minteq.v4.dat, eqn = "Pb+2 + Cl- = PbCl+", logk = 1.59)[1820:1821]
}
