% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cu_sol.R
\name{cu_sol}
\alias{cu_sol}
\title{Calculate equilibrium copper solubility on a grid of input values}
\usage{
cu_sol(
  ph,
  dic,
  chloride = 0,
  sulfate = 0,
  phosphate = 0,
  phase,
  db = cu2sol,
  empirical = FALSE
)
}
\arguments{
\item{ph}{A vector of pH values.}

\item{dic}{A vector of dissolved inorganic carbon concentrations, in mg C/L.}

\item{chloride}{A vector of chloride concentrations, in  mg/L.}

\item{sulfate}{A vector of sulfate concentrations, in mg SO4/L.}

\item{phosphate}{A vector of orthophosphate concentrations, in mg P/L.}

\item{phase}{Equilibrium phase.}

\item{db}{The database used in equilibrium solubility computations. The default is \code{pbcusol:::cu2sol}}

\item{empirical}{Logical. Predict equilibrium copper solubility using the empirical model due to Lytle et al.?
(https://doi.org/10.1002/awwa.1109)}
}
\value{
A tibble with the input grid of pH and DIC values and the copper solubilty predictions.
}
\description{
Calculate equilibrium copper solubility on a grid of input values
}
\examples{
cu_sol(ph = 7.5, dic = 5, phase = "Tenorite")
cu_sol(ph = 7.5, dic = 5, db = phreeqc::minteq.v4.dat, phase = "Tenorite")
cu_sol(7.5, 5, phosphate = 1, empirical = TRUE)
}
