% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq_sol_fixed.R
\name{eq_sol_fixed}
\alias{eq_sol_fixed}
\alias{pb_sol_fixed}
\alias{cu_sol_fixed}
\title{Calculate equilibrium solubility with fixed pH and pe}
\usage{
eq_sol_fixed(
  ph,
  dic,
  phosphate = 0,
  phase,
  element,
  phase_quantity = 1,
  pe = 4,
  eq_phase_components = list(),
  new_phase = list(),
  phase_out = "Fix_pH",
  new_species = list(),
  surface_components = list(),
  output_components = list(),
  buffer = "NaOH",
  db = pbcu2sol,
  print = NULL,
  ...
)

pb_sol_fixed(..., element = "Pb")

cu_sol_fixed(..., element = "Cu")
}
\arguments{
\item{ph}{pH}

\item{dic}{Dissolved inorganic carbon, in mg C/L.}

\item{phosphate}{Orthophosphate, in mg P/L.}

\item{phase}{Equilibrium phase.}

\item{element}{An element to return the equilibrium concentration of.}

\item{phase_quantity}{Moles of equilibrium phase initially present.}

\item{pe}{Set the pe of the solution.}

\item{eq_phase_components}{Additional equilibrium phase components, passed to
\code{tidyphreeqc::phr_input_section} as a list.}

\item{new_phase}{Define phases not included in the database.}

\item{phase_out}{Add an equilibrium phase to the output. Default is the pseudophase "Fix_pH".}

\item{new_species}{Define solution species not included in the database.}

\item{surface_components}{Components of a surface assemblage, passed to
\code{tidyphreeqc::phr_input_section} as a list.}

\item{output_components}{Additional output components, passed to
\code{tidyphreeqc::phr_input_section} as a list.}

\item{buffer}{Substance added or subtracted from the solution to yield the desired pH.}

\item{db}{The database to use for equilibrium solubility computations. The default is
\code{pbcusol:::leadsol}}

\item{print}{Choose whether to print the input file ("input"), the full output ("output"), or the selected output.
Default is the latter.}

\item{...}{Arguments passed on to \code{tidyphreeqc::phr_input_section()} as solution phase
components. Concentrations should be expressed in mmol/kgw.}
}
\value{
A tibble with columns representing equilibrium phase, pH, dissolved inorganic carbon,
orthophosphate (as P), pe, ionic strength (mu), total concentration of chosen element in solution, and moles of the equilibrium phase
dissolved.
}
\description{
Simulating extreme conditions may require fixing the pH or pe. This is
accomplished here using \code{tidyphreeqc::phr_pH_fix_definition()}, with help from
\url{https://swilke-geoscience.net/post/phreeqc-mineral-solubility/}.
}
\section{Functions}{
\itemize{
\item \code{pb_sol_fixed}: Shorthand for \code{eq_sol_fixed()} with \code{element = "Pb"}. For backwards compatibility.

\item \code{cu_sol_fixed}: Shorthand for \code{eq_sol_fixed()} with \code{element = "Cu"}. For backwards compatibility.
}}

\examples{
eq_sol_fixed(element = "Pb", ph = 6, dic = 5, phase = "Cerussite", buffer = "HCl")
}
