% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pb_logk.R
\name{pb_logk}
\alias{pb_logk}
\title{Return LEADSOL phases, reactions, and log K values}
\usage{
pb_logk(kable_format = FALSE)
}
\arguments{
\item{kable_format}{Logical. Format the table for \code{knitr::kable}?}
}
\value{
A tibble with phase names, reactions, and log K values.
}
\description{
Return LEADSOL phases, reactions, and log K values
}
\examples{
pb_logk()
}
