% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pb_sol.R
\name{pb_sol}
\alias{pb_sol}
\title{Calculate equilibrium lead solubility on a grid of input values}
\usage{
pb_sol(ph, dic, chloride = 0, sulfate = 0, phosphate = 0, phase, db = leadsol)
}
\arguments{
\item{ph}{A vector of pH values.}

\item{dic}{A vector of dissolved inorganic carbon (DIC) concentrations, in mg C/L.}

\item{chloride}{A vector of chloride concentrations, in  mg/L.}

\item{sulfate}{A vector of sulfate concentrations, in mg SO4/L.}

\item{phosphate}{A vector of orthophosphate concentrations, in mg P/L.}

\item{phase}{Equilibrium phase.}

\item{db}{The database to use for equilibrium solubility computations. The default is \code{pbcusol:::leadsol}}
}
\value{
A tibble with the input grid of pH and DIC values and the lead solubilty predictions.
}
\description{
Calculate equilibrium lead solubility on a grid of input values
}
\examples{
pb_sol(ph = 7.5, dic = 5, phase = "Cerussite")
pb_sol(ph = 7.5, dic = 5, db = phreeqc::minteq.v4.dat, phase = "Cerussite")
}
