% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pb_sol_fixed.R
\name{pb_sol_fixed}
\alias{pb_sol_fixed}
\title{Calculate equilibrium solubility with fixed pH}
\usage{
pb_sol_fixed(
  ph,
  dic,
  phosphate = 0,
  phase,
  element = "Pb",
  phase_quantity = 10,
  eq_phase_components = list(),
  new_phase = list(),
  phase_out = "Fix_pH",
  new_species = list(),
  surface_components = list(),
  output_components = list(),
  buffer = "NaOH",
  db = pbcusol:::leadsol,
  ...
)
}
\arguments{
\item{ph}{pH}

\item{dic}{Dissolved inorganic carbon, in mg C/L.}

\item{phosphate}{Orthophosphate, in mg P/L.}

\item{phase}{Equilibrium phase.}

\item{element}{An element to return the equilibrium concentration of. Default is Pb.}

\item{phase_quantity}{Moles of equilibrium phase initially present.}

\item{eq_phase_components}{Additional equilibrium phase components, passed to
\code{tidyphreeqc::phr_input_section} as a list.}

\item{new_phase}{Define phases not included in the database.}

\item{phase_out}{Add an equilibrium phase to the output. Default is the pseudophase "Fix_pH".}

\item{new_species}{Define solution species not included in the database.}

\item{surface_components}{Components of a surface assemblage, passed to
\code{tidyphreeqc::phr_input_section} as a list.}

\item{output_components}{Additional output components, passed to
\code{tidyphreeqc::phr_input_section} as a list.}

\item{buffer}{Substance added or subtracted from the solution to yield the desired pH.}

\item{db}{The database to use for equilibrium solubility computations. The default is
\code{pbcusol:::leadsol}}

\item{...}{Arguments passed on to \code{tidyphreeqc::phr_input_section()} as solution phase
components. Concentrations should be expressed in mmol/kgw.}
}
\value{
A tibble with columns representing equilibrium phase, pH, dissolved inorganic carbon,
orthophosphate (as P), pe, ionic strength (mu), total lead, and moles of the equilibrium phase
dissolved.
}
\description{
While most lead phases are not soluble enough to have a substantial effect on
pH (or pe) as they dissolve, simulating extreme conditions may require fixing the pH. This is
accomplished here using \code{tidyphreeqc::phr_pH_fix_definition()}, with help from
\url{https://swilke-geoscience.net/post/phreeqc-mineral-solubility/}.
}
\examples{
pb_sol_fixed(ph = 6, dic = 5, phase = "Cerussite", buffer = "HCl")
}
