% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pb_sol_jurgens.R
\name{pb_sol_jurgens}
\alias{pb_sol_jurgens}
\title{Calculate equilibrium lead solubility, approximating the method of Jurgens et al.}
\usage{
pb_sol_jurgens(
  Pb = 1,
  ph,
  dic,
  phosphate = 0,
  phase,
  element = "Pb",
  eq_phase_components = list(),
  new_phase = list(),
  phase_out = "Fix_pH",
  new_species = list(),
  surface_components = list(),
  output_components = list(),
  buffer = "NaOH",
  db = phreeqc::Amm.dat,
  print = NULL,
  ...
)
}
\arguments{
\item{Pb}{Concentrations of Pb initially present in solution (micrograms/L).}

\item{ph}{pH}

\item{dic}{Dissolved inorganic carbon, in mg C/L.}

\item{phosphate}{Orthophosphate, in mg P/L.}

\item{phase}{Equilibrium phase.}

\item{element}{An element to return the equilibrium concentration of.}

\item{eq_phase_components}{Additional equilibrium phase components, passed to
\code{tidyphreeqc::phr_input_section} as a list.}

\item{new_phase}{Define phases not included in the database.}

\item{phase_out}{Add an equilibrium phase to the output. Default is the pseudophase "Fix_pH".}

\item{new_species}{Define solution species not included in the database.}

\item{surface_components}{Components of a surface assemblage, passed to
\code{tidyphreeqc::phr_input_section} as a list.}

\item{output_components}{Additional output components, passed to
\code{tidyphreeqc::phr_input_section} as a list.}

\item{buffer}{Substance added or subtracted from the solution to yield the desired pH.}

\item{db}{The database to use for equilibrium solubility computations. The default is
\code{pbcusol:::leadsol}}

\item{print}{Choose whether to print the input file ("input"), the full output ("output"), or the selected output.
Default is the latter.}

\item{...}{Arguments passed on to \code{tidyphreeqc::phr_input_section()} as solution phase
components. Concentrations should be expressed in mmol/kgw.}
}
\value{
A tibble with columns representing equilibrium phase, pH, dissolved inorganic carbon,
orthophosphate (as P), pe, ionic strength (mu), total concentration of chosen element in solution, and moles of the equilibrium phase
dissolved.
}
\description{
Experimental implementation of the approach to lead solubility calculation outlined in Jurgens et al.
\url{https://doi.org/10.1021/acs.est.8b04475}.
}
\examples{
pb_sol_jurgens(ph = 8, dic = 5)
}
