% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pb_sol_wham.R
\name{pb_sol_wham}
\alias{pb_sol_wham}
\title{Calculate equilibrium lead solubility in the presence of humic acid}
\usage{
pb_sol_wham(
  ph,
  dic,
  phosphate = 0,
  phase,
  mass_ha = 0,
  mu_is = 0.003,
  eq_phase_components = list(),
  new_phase = list(),
  phase_out = "Fix_pH",
  new_species = list(),
  surface_components = list(),
  output_components = list(`-totals` = c("P", "C", "Pb")),
  buffer = "NaOH",
  db = pbcusol:::leadsol,
  max_iter = 3,
  ...
)
}
\arguments{
\item{ph}{pH}

\item{dic}{Dissolved inorganic carbon, in mg C/L.}

\item{phosphate}{Orthophosphate, in mg P/L.}

\item{phase}{Equilibrium phase.}

\item{mass_ha}{Mass of humic acid in grams dissolved organic carbon.}

\item{mu_is}{An initial guess for the ionic strength, used to estimate the specific surface
area of the humic acid molecules.}

\item{eq_phase_components}{Additional equilibrium phase components, passed to
\code{tidyphreeqc::phr_input_section} as a list.}

\item{new_phase}{Define phases not included in the database.}

\item{phase_out}{Add an equilibrium phase to the output. Default is the pseudophase "Fix_pH".}

\item{new_species}{Define solution species not included in the database.}

\item{surface_components}{Additional surface components, passed to
\code{tidyphreeqc::phr_input_section} as a list.}

\item{output_components}{Additional output components, passed to
\code{tidyphreeqc::phr_input_section} as a list.}

\item{buffer}{Substance added or subtracted from the solution to yield the desired pH.}

\item{db}{The database to use for equilibrium solubility computations. The default is
\code{pbcusol:::leadsol}}

\item{max_iter}{Maximum iterations allowed for convergence of ionic strength.}

\item{...}{Arguments passed on to \code{tidyphreeqc::phr_input_section()} as solution phase
components. Concentrations should be expressed in mmol/kgw.}
}
\value{
A tibble with columns representing equilibrium phase, pH, dissolved inorganic carbon,
orthophosphate (as P), pe, ionic strength (mu), total lead in solution, and moles of the equilibrium phase
dissolved.
}
\description{
Compute equilibrium lead solubility in the presence of humic acid, using the
Windermere Humic Acid Model (WHAM), due to Tipping and Hurley (see \code{phreeqc::Tipping_Hurley.dat},
\url{https://doi.org/10.1016/0016-7037(92)90158-F}, and
\url{https://water.usgs.gov/water-resources/software/PHREEQC/documentation/phreeqc3-html/phreeqc3.htm}).
This function works like \code{pb_sol_fixed()}.
}
\examples{
pb_sol_wham(ph = 7.5, dic = 50, phase = "Cerussite", Na = 10, mass_ha = 3.5e-3)
}
