% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cu_logk.R
\name{cu_logk}
\alias{cu_logk}
\title{Return CU2SOL phases, reactions, and log K values}
\usage{
cu_logk(kable_format = FALSE)
}
\arguments{
\item{kable_format}{Logical. Format the table for \code{knitr::kable}?}
}
\value{
A tibble with phase names, reactions, and log K values.
}
\description{
Return CU2SOL phases, reactions, and log K values
}
\examples{
cu_logk()
}
