% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cu_sol.R
\name{cu_sol}
\alias{cu_sol}
\title{Calculate equilibrium copper solubility on a grid of input values}
\usage{
cu_sol(ph, dic, chloride = 0, sulfate = 0, phosphate = 0, phase, db = cu2sol)
}
\arguments{
\item{ph}{A vector of pH values.}

\item{dic}{A vector of dissolved inorganic carbon concentrations, in mg C/L.}

\item{chloride}{A vector of chloride concentrations, in  mg/L.}

\item{sulfate}{A vector of sulfate concentrations, in mg SO4/L.}

\item{phosphate}{A vector of orthophosphate concentrations, in mg P/L.}

\item{phase}{Equilibrium phase.}

\item{db}{The database used in equilibrium solubility computations. Default is \code{pbcusol:::cu2sol}}
}
\value{
A numeric vector with lead solubilty predictions at the input settings.
}
\description{
Calculate equilibrium copper solubility on a grid of input values
}
\examples{
cu_sol(ph = 7.5, dic = 5, phase = "Tenorite")
cu_sol(ph = 7.5, dic = 5, db = phreeqc::minteq.v4.dat, phase = "Tenorite")
}
